refactor: simplify snapshot, remove detail namespace and snapshot.hpp files
