+++ src/module/predictor/outpost/snapshot.cpp
@@ -1,4 +1,6 @@
-#include "module/predictor/outpost/snapshot.hpp"
+#include "module/predictor/outpost/armor_layout.hpp"
+#include "module/predictor/snapshot.hpp"
+#include "utility/robot/color.hpp"
 
 #include <algorithm>
 #include <utility>
@@ -11,9 +13,8 @@
 namespace rmcs::predictor {
 
 namespace {
-
     auto make_armor(DeviceId device, CampColor color, int id) -> Armor3D {
-        auto armor  = Armor3D { };
+        auto armor  = Armor3D {};
         armor.genre = device;
         armor.color = camp_color2armor_color(color);
         armor.id    = id;
@@ -34,7 +35,7 @@ namespace {
         [[nodiscard]] auto predicted_armors(TimePoint t) const -> std::vector<Armor3D> override {
             auto const predicted_x = predict_state_at(t);
 
-            auto armors = std::vector<Armor3D> { };
+            auto armors = std::vector<Armor3D> {};
             armors.reserve(std::clamp(armor_num, 0, OutpostEKFParameters::kOutpostArmorCount));
 
             for (int id = 0; id < OutpostEKFParameters::kOutpostArmorCount; ++id) {
@@ -79,10 +80,10 @@ namespace {
 
 } // namespace
 
-auto detail::make_outpost_snapshot(Snapshot::OutpostEKF::XVec ekf_x, CampColor color, int armor_num,
+auto make_outpost_snapshot(Snapshot::OutpostEKF::XVec ekf_x, CampColor color, int armor_num,
     TimePoint stamp, OutpostArmorLayout outpost_layout) noexcept -> Snapshot {
-    return detail::make_snapshot(std::make_unique<OutpostSnapshotBackend>(
-        std::move(ekf_x), color, armor_num, stamp, outpost_layout));
+    return Snapshot { std::make_unique<OutpostSnapshotBackend>(
+        std::move(ekf_x), color, armor_num, stamp, outpost_layout) };
 }
 
 } // namespace rmcs::predictor
