+++ src/module/predictor/snapshot.hpp
@@ -6,17 +6,13 @@
 
 #include "utility/clock.hpp"
 #include "utility/math/kalman_filter/ekf.hpp"
-#include "utility/robot/id.hpp"
 #include "utility/robot/armor.hpp"
+#include "utility/robot/id.hpp"
 
 namespace rmcs::predictor {
 
 struct ISnapshotBackend;
 class Snapshot;
-
-namespace detail {
-    auto make_snapshot(std::unique_ptr<ISnapshotBackend> backend) noexcept -> Snapshot;
-}
 
 class Snapshot {
 public:
@@ -28,30 +24,25 @@ public:
         double angular_velocity;
     };
 
-    static auto empty(TimePoint stamp) noexcept -> Snapshot;
-
+    explicit Snapshot(std::unique_ptr<ISnapshotBackend> backend) noexcept;
     Snapshot(Snapshot const&) = delete;
     Snapshot(Snapshot&&) noexcept;
     Snapshot& operator=(Snapshot const&) = delete;
     Snapshot& operator=(Snapshot&&) noexcept;
     ~Snapshot() noexcept;
+    static auto empty(TimePoint stamp) noexcept -> Snapshot;
 
     auto time_stamp() const -> TimePoint;
     auto device_id() const -> DeviceId;
+
     auto kinematics() const -> Kinematics;
     auto kinematics_at(TimePoint t) const -> Kinematics;
 
     auto predicted_armors() const { return predicted_armors(Clock::now()); }
-
     auto predicted_armors(TimePoint t) const -> std::vector<Armor3D>;
 
 private:
-    explicit Snapshot(std::unique_ptr<ISnapshotBackend> backend) noexcept;
-
     std::unique_ptr<ISnapshotBackend> backend;
-
-    friend auto detail::make_snapshot(std::unique_ptr<ISnapshotBackend> backend) noexcept
-        -> Snapshot;
 };
 
 }
